+++ organization/src/epilot/organization.py
@@ -20,7 +20,7 @@ class Organization:
         url = utils.generate_url(operations.GetOrganizationRequest, base_url, '/v2/organization/{org_id}', request)
         headers = {}
         headers['Accept'] = 'application/json'
-        headers['user-agent'] = f'speakeasy-sdk/{self.sdk_configuration.language} {self.sdk_configuration.sdk_version} {self.sdk_configuration.gen_version} {self.sdk_configuration.openapi_doc_version}'
+        headers['user-agent'] = self.sdk_configuration.user_agent
         
         client = self.sdk_configuration.security_client
         
@@ -45,11 +45,11 @@ class Organization:
         
         url = utils.generate_url(operations.UpdateOrganizationRequest, base_url, '/v2/organization/{org_id}', request)
         headers = {}
-        req_content_type, data, form = utils.serialize_request_body(request, "organization", 'json')
+        req_content_type, data, form = utils.serialize_request_body(request, "organization", False, True, 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
         headers['Accept'] = 'application/json'
-        headers['user-agent'] = f'speakeasy-sdk/{self.sdk_configuration.language} {self.sdk_configuration.sdk_version} {self.sdk_configuration.gen_version} {self.sdk_configuration.openapi_doc_version}'
+        headers['user-agent'] = self.sdk_configuration.user_agent
         
         client = self.sdk_configuration.security_client